add better script
