--- docs/dev-guide/overview/index.md
@@ -1,3 +1,0 @@
-# Project overview
-
-These pages are designed to orient and find certain areas of the PrarieLearn repository. The navigational structure mirrors the folder structure of PrairieLearn. You can click the **Edit this page :pencil:** link while viewing them to jump the source file on GitHub.
