+++ mkdocs.yml
@@ -69,7 +69,7 @@ plugins:
   - gen-files:
       scripts:
         - docs/scripts/gen_python_nav.py
-        - docs/scripts/gen_developer_nav.py
+        - docs/scripts/gen_addl_readmes.py
   # Used for documenting our code with an autogenerated markdown file
   - literate-nav:
       implicit_index: true
@@ -181,6 +181,10 @@ extra_css:
   - assets/extra.css
 
 extra:
+  # This section is consumed by the gen_addl_readmes.py script
+  addl_readmes:
+    - src: './contrib'
+      dest: 'dev-guide/contrib/'
   social:
     - icon: fontawesome/brands/github
       link: https://github.com/PrairieLearn/PrairieLearn
@@ -235,7 +239,7 @@ nav:
       - 'Building element extensions': 'elementExtensions.md'
       - 'Code execution': 'codeExecution.md'
       - 'Contributing': 'contributing.md'
-      - 'Project overview': 'dev-guide/overview/'
+      - 'Utility scripts': 'dev-guide/contrib/'
   - Administrators:
       - 'SAML SSO configuration': 'saml.md'
       - 'LTI 1.3 configuration': 'lti13.md'
